* Don't attempt to list a non-existent directory (Bug #32813)
git-svn-id: svn+ssh://billy.knut.univention.de/var/univention/svn/dev/branches/ucs-4.0/ucs-4.0-0@53013 1b283449-9f2f-0410-b571-8eb228e1a901

+++ services/univention-s4-connector/s4-connector.py
@@ -166,6 +166,8 @@ def clean():
 	listener.setuid(0)
 	try:
 		for directory in dirs:
+			if not os.path.exists(directory):
+				continue
 			for filename in os.listdir(directory):
 				if filename != "tmp":
 					os.remove(os.path.join(directory,filename))
